Upload image breaking issue fixed
This issue comes after updating to the latest version of element-plus. It breaks the svg image if user upload a application type file. As element plus's css file getting more prority on after loading so we have to use !important for this. Also this is a part of our core application so user won't change this in future, and for that we can use the !important for this to resolve the issue.