Fix editor vanish issue in customer end
This issue was occurring when the customer reloads the ticket create page. On page load js var(window.wp) isn't loading, as wp_editor not getting this value on hasEditor property and in editor property so it throws an error on the promise, and therefore customers' ticket creation page failed to load after reloading the page.