Added ticket migration from JS Helpdesk
It added initially and all of the parts of this module working perfectly. We have to improve the deleting system of js helpdesk after migration. resolvePerson method moved to baseimporter as it's used by more than one module maybe we will get more like this later. And added ticket_id on reply attachment it won't create any issues if we don't do this as attachments of replies fetch using conversattion id.