Ticket filtering issue fixed
Ticket filter get reset when someone go to a ticket page and comes again in the tickets page or if someone click on the browser back button. Even if someone browse another page like report, dashboard or any other page then if that person come again in the tickets page then he lost the saved filter.